Fait: Ajouter style a la barre et ajouter une barre de recherche aux offres de stages
