+++ frontend/src/Components/Form/SearchInput.jsx
@@ -1,0 +1,22 @@
+export default function SearchInput({ searchIndex, setSearchIndex }) {
+
+    const handleSearchInputChange = (event) => {
+        event.preventDefault();
+        let value = event.target.value;
+        setSearchIndex(value);
+    }
+
+    return (
+        <form className="justify-center">
+            <label for="default-search" class="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white">Search</label>
+            <div class="relative">
+                <div class="absolute inset-y-0 left-0 flex items-center pl-3 pointer-events-none">
+                    <svg class="w-4 h-4 text-gray-500 dark:text-gray-400" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
+                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z" />
+                    </svg>
+                </div>
+                <input onChange={handleSearchInputChange} type="search" id="default-search" class="block w-full p-4 pl-10 text-sm text-black border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500" placeholder="Rechercher un stage..." required/>
+            </div>
+        </form>
+    )
+}
